+++ uni_modules/uni-id/changelog.md
@@ -1,3 +1,7 @@
+## 3.1.3（2021-07-08）
+- 移除插件内误传的node_modules
+## 3.1.2（2021-07-08）
+- 修复 微信小程序绑定微信账号时报错的Bug
 ## 3.1.1（2021-07-01）
 - 使用新的错误码规范，兼容旧版 [详情](https://uniapp.dcloud.net.cn/uniCloud/uni-id?id=errcode)
 - 修复微信登录、绑定时未返回用户accessToken的Bug